fixed warning

+++ ugbase/common/util/tar.h
@@ -47,7 +47,7 @@ struct TarHeader
 		strcpy(ustarDeviceMajorNumber, "000000 ");
 		strcpy(ustarDeviceMinorNumber, "000000 ");
 
-		sprintf(octalModificationTimeStamp, "%o", (unsigned)time(NULL));
+		sprintf(octalModificationTimeStamp, "%o", ((unsigned int)time(NULL)));
 //		strncpy(octalModificationTimeStamp, "12253557334 ", 12);
 		linkIndicator = '0';
 	}
@@ -58,7 +58,7 @@ struct TarHeader
 	void set_filesize(size_t size)
 	{
 		char buf[13];
-		sprintf(buf, "%011o ", size);
+		sprintf(buf, "%011o ", (unsigned int)size);
 		memcpy(octalFileSize, buf, 12);
 	}
 
@@ -66,10 +66,10 @@ struct TarHeader
 	{
 		memset(checksum, 32, sizeof(checksum));
 		unsigned char *p = (unsigned char *) this;
-		unsigned long cs = 0;
+		unsigned int cs = 0;
 		for(size_t i=0; i<sizeof(TarHeader); i++)
 			cs += p[i];
-		sprintf(checksum, "%06o", cs);
+		sprintf(checksum, "%06o", ((unsigned int)cs));
 		//			 std:: cout << "checksum = " << checksum << "\n";
 	}
 };
